sqli-demo-be/requirements.txt
Flask==2.2.2  
Werkzeug==2.2.2  
Flask-CORS==3.0.10
mysql-connector-python==8.0.33
SQLAlchemy==2.0.20